+++ example/container_tile.py
@@ -1,5 +1,5 @@
 import cadquery as cq
-from cqportal import FloorTile, Container
+from cqportal import FloorTile, Container, RampGreebled
 from cqterrain import tile as terrain_tile
 
 #----------------------
@@ -15,10 +15,15 @@ def make_custom_tile(length, width, height):
     )
 
 bp_container = Container()
-bp_container.render_ladder = True
+bp_container.render_ladder = False
 #bp_container.bp_ladder.ladder_rungs = 6
 #bp_container.bp_ladder.ladder_depth = 5
 #bp_container.bp_ladder.ladder_rung_radius = 1.5
+
+bp_container.bp_ramp = RampGreebled()
+bp_container.bp_ramp.width = 8
+bp_container.bp_ramp.segment_count = 10
+bp_container.bp_ramp.segment_y_padding = 3
 
 bp_container.bp_floor = FloorTile()