+++ src/cqportal/shieldwall/__init__.py
@@ -1,0 +1,2 @@
+from .ShieldShape import ShieldShape, shield_shape, shield_shape_mirror
+from .StraightBasic import StraightBasic