+++ changes.md
@@ -1,6 +1,12 @@
 # CQPortal changelog
 
 ## Main wip
+## 2.1.0
+* synced container tile example.
+* Added initial shieldwall ShieldShape.
+* Added initial shieldwall StraightBasic
+* Added stubs for new Shield wall shapes
+
 ## 2.0.0
 * Refactored Ramp to be an orchestrator
 * Added BaseCoffin